# lib/api-spec/openapi_half.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PMFBY Crop Insurance AI Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /farmers:
    get:
      operationId: listFarmers
      tags: [farmers]
      summary: List all farmers
      parameters:
        - name: district
          in: query
          schema: { type: string }
        - name: state
          in: query
          schema: { type: string }
        - name: search
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: List of farmers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Farmer"
    post:
      operationId: createFarmer
      tags: [farmers]
      summary: Register a new farmer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FarmerInput"
      responses:
        "201":
          description: Farmer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Farmer"
  /farmers/{id}:
    get:
      operationId: getFarmer
      tags: [farmers]
      summary: Get farmer details
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Farmer details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Farmer"
        "404":
          description: Not found
    patch:
      operationId: updateFarmer
      tags: [farmers]
      summary: Update farmer details
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FarmerUpdate"
      responses:
        "200":
          description: Updated farmer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Farmer"
  /farmers/{id}/claims:
    get:
      operationId: getFarmerClaims
      tags: [farmers]
      summary: Get all claims for a farmer
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Claims for farmer
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Claim"

  # ─── Fields ──────────────────────────────────────────────────────────────
  /fields:
    get:
      operationId: listFields
      tags: [fields]
      summary: List agricultural fields
      parameters:
        - name: farmerId
          in: query
          schema: { type: integer }
        - name: district
          in: query
          schema: { type: string }
        - name: cropType
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: List of fields
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Field"
    post:
      operationId: createField
      tags: [fields]
      summary: Register a new field
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FieldInput"
      responses:
        "201":
          description: Field created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Field"
  /fields/{id}:
    get:
      operationId: getField
      tags: [fields]
      summary: Get field details
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Field details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Field"
        "404":
          description: Not found

  # ─── Claims ──────────────────────────────────────────────────────────────
...
  /inspections:
    get:
      operationId: listInspections
      tags: [inspections]
      summary: List field staff inspections
      parameters:
        - name: claimId
          in: query
          schema: { type: integer }
        - name: staffId
          in: query
          schema: { type: integer }
        - name: status
          in: query
          schema:
            type: string
            enum: [scheduled, in_progress, completed, cancelled]
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: List of inspections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Inspection"
    post:
      operationId: createInspection
      tags: [inspections]
      summary: Schedule a field inspection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InspectionInput"
      responses:
        "201":
          description: Inspection scheduled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inspection"
  /inspections/{id}:
    get:
      operationId: getInspection
      tags: [inspections]
      summary: Get inspection details
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Inspection details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inspection"
        "404":
          description: Not found
    patch:
      operationId: updateInspection
      tags: [inspections]
      summary: Update inspection findings
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InspectionUpdate"
      responses:
        "200":
          description: Updated inspection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inspection"

  # ─── Anomalies ────────────────────────────────────────────────────────────
...